♻️ #219 - 최근 검색어 컨테이너 분리
