+++ src/components/search/index.ts
@@ -1,1 +1,3 @@
 export * from './SearchHeader';
+export * from './NoSearchResults';
+export * from './RecentSearchContainer';
